requirements.txt
backtrader~=1.9.76.123
matplotlib~=3.2.2
pyyaml~=6.0
akshare~=1.7.54
pandas~=1.5.0